Update webpack.config for new MAPBOX_TOKEN
We are using an environment variable of MAPBOX_TOKEN to keep track of
our tokens. The dotenv package will ensure that if the token is set in
the .env file it will be available on the process.

However, if that keyword is set on the process already it will not be
overwritten.